# server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Loventia API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Loventia API Team
    url: https://loventia.example.com
    email: api@loventia.example.com
  description: |
    OpenAPI specification for the Loventia dating application.

    This spec is aligned with the current server/src/app.js routing we just test-drove
    with PowerShell:

      - Primary auth endpoint: POST /api/auth/login
      - Legacy/fallback auth endpoint: POST /api/users/login
      - Refresh: POST /api/auth/refresh
      - Self: GET /api/auth/me and GET /api/users/me and GET /api/me
      - Reset-password flow: POST /api/auth/forgot-password, POST /api/auth/reset-password
      - Private diagnostics: GET /api/auth/private/ping, GET /api/auth/private/no-me
      - User photo management (the things we just tested): 
          * DELETE /api/users/{userId}/photos/{slot}
          * DELETE /api/users/{userId}/photos?path=...
          * PUT    /api/users/{userId}/photos/reorder
          * POST   /api/users/{userId}/set-avatar
      - Billing endpoints (Stripe webhook at /webhooks/stripe and legacy /api/payment/stripe-webhook)
      - Admin CSV exports
      - Health/readiness endpoints

    IMPORTANT:
      - We **document both** logins on purpose:
          1) /api/auth/login  (main, preferred)
          2) /api/users/login (legacy/fallback)
      - This matches the README change you will add in server/README.md
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: villehermaala1981@gmail.com
        password:
          type: string
          example: Paavali1981
    AuthLoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful.
        user:
          $ref: '#/components/schemas/UserSelf'
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          example: 7200
        # OpenAPI 3.1: use union type instead of nullable
        refreshExpiresIn:
          type: [integer, 'null']
    AuthRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/UserSelf'
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: testuser1@example.com
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          description: Reset token from email link
        password:
          type: string
          format: password
          minLength: 6
        id:
          type: string
          description: Optional user ID if backend expects it
    UserSelf:
      type: object
      description: >
        Minimal user shape returned by /api/auth/me and /api/users/me in our PS tests.
        Includes entitlements, visibility and premium flags.
      properties:
        id:
          type: string
          example: 68dab8a761a412065159ff4d
        _id:
          type: string
          example: 68dab8a761a412065159ff4d
        email:
          type: string
          example: villehermaala1981@gmail.com
        username:
          type: string
          example: Ville1981
        role:
          type: string
          example: user
        premium:
          type: boolean
          example: true
        isPremium:
          type: boolean
          example: true
        entitlements:
          type: object
          properties:
            tier:
              type: string
              example: premium
            # OpenAPI 3.1 union types (replace nullable)
            since:
              type: [string, 'null']
            until:
              type: [string, 'null']
            features:
              type: object
              additionalProperties: true
            quotas:
              type: object
              additionalProperties: true
        visibility:
          type: object
          properties:
            isHidden:
              type: boolean
              example: false
            # OpenAPI 3.1 union type (replace nullable)
            hiddenUntil:
              type: [string, 'null']
            resumeOnLogin:
              type: boolean
              example: true
        profilePicture:
          type: string
          example: /uploads/1759343842730-848316731.gif
        photos:
          type: array
          items:
            type: string
            example: /uploads/1759343842730-848316731.gif
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login (primary)
      description: >
        Primary login endpoint for Loventia. FE calls this FIRST.
        If it returns 404/405, FE may try /api/users/login (legacy).
      operationId: auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: Invalid credentials or missing payload
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: >
        Matches the PS call: POST /api/auth/refresh { "refreshToken": "<token>" }
        Returns a fresh access token and user payload.
      operationId: auth_refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRefreshResponse'
        '401':
          description: Invalid or expired refresh token
...
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout (stateless)
      description: >
        Stateless logout. PowerShell test showed:
        { "message": "Logout successful" }
        and existing access token still worked (expected with stateless).
      operationId: auth_logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
        '401':
          description: Unauthorized
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset email
      description: Send a reset email to the provided address.
      operationId: auth_forgot_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent (or noop for security)
        '400':
          description: Invalid email
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password with token
      description: >
        Matches FE flow: POST /api/auth/reset-password { token, password, id? }
        BE responds "Password has been reset successfully."
      operationId: auth_reset_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password has been reset successfully.
        '400':
          description: Invalid or expired token
...
  /__routes:
    get:
      tags: [Diagnostics]
      summary: List app-level routes (dev-only)
      description: Returns app-level mounted routes (dev-only).
      operationId: diagnostics_routes
      responses:
        '200':
          description: List of routes
        '401':
          description: Unauthorized
...
security:
  - bearerAuth: []